Added support for a queue object
A queue is a specific object that implements a FIFO queue of some other element.

Type code is now split up into object and type parts.
It is also now possible for an object to be multiple structures (but not multiple typedefs)

Also, added a placeholder for a BST object.